requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
drf-spectacular==0.27.1
python-dotenv==1.0.1
google-generativeai==0.4.1
Pillow==10.2.0
psycopg[binary]==3.3.2